Phase 1: split TrainlogService into per-domain API classes
Break the 957-line TrainlogService monolith into a shared
TrainlogHttpClient (Dio + cookie jar + safe request helpers) and six
domain APIs under lib/services/api/:

- trainlog_http_client.dart: transport, baseUrl, instance constants,
  get/post passthroughs, safeGet/safeGetWithRetry/safePost
- auth_api.dart: login, signup, checkAuthenticated, clearSession,
  fetchUsernameViaApi, getCookiesForWebView, TrainlogLoginResult
- trips_api.dart: fetchAllTripsData, fetchLastUpdatedTripsData,
  deleteTrip(s)
- stations_api.dart: fetchStations, manual stations, find stations
  from coordinate, list generators, StationResult
- account_api.dart: account settings + available currencies
- misc_api.dart: stats, visited stations, operator logos, news

TrainlogProvider now owns the client and builds the domain APIs, keeping
its public surface unchanged. TripsProvider takes a TripsApi; the WebView
shells call provider.getCookiesForWebView() directly.
